Refactor code base folder structure. Other changes
