+++ src/components/icons/DocumentTypeIcon.tsx
