+++ src/layouts/ContentLayout.tsx
@@ -1,5 +1,5 @@
 import { useLocation } from "react-router-dom";
-import Header from "../components/Header";
+import Header from "../components/common/Header";
 import { useSidebar } from "../hooks/UseSidebar";
 
 type ContentLayoutProps = {
