+++ src/pages/KnowledgeBase.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import SearchAndFilter from "../components/SearchAndFilter";
-import VideoCard from "../components/VideoCard";
+import SearchAndFilter from "../components/common/SearchAndFilter";
+import VideoCard from "../components/cards/VideoCard";
 import { Outlet, useLocation, useNavigate } from "react-router-dom";
 import { videoData } from "../mocks/AllMockData";
 import { VideoDataTypes } from "../types/AllTypes";
