+++ src/pages/auth/Login.tsx
@@ -2,10 +2,10 @@ import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import Forms from "../../components/auth/Forms";
 import { logInSchema } from "../../validation/schema";
-import FormInput from "../../components/auth/FormInput";
+import FormInput from "../../components/form-components/FormInput";
 import { ILoginForm } from "../../types/AllTypes";
 import { NavLink, useNavigate } from "react-router-dom";
-import CustomButton from "../../components/CustomButton";
+import CustomButton from "../../components/form-components/CustomButton";
 import { useState } from "react";
 import { ClipLoader } from "react-spinners";
